odop-backend/pyproject.toml
[project]
name = "odop-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "ruff>=0.13.1",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "pytest-django>=4.11.1",
]
